refactor: di 모듈의 패키지 이름 변경
