+++ woogi-di/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     kotlin("jvm")
+    id("maven-publish")
 }
 
 java {
@@ -18,5 +19,16 @@ dependencies {
 tasks {
     test {
         useJUnitPlatform()
+    }
+}
+
+publishing {
+    publications {
+        create<MavenPublication>("maven") {
+            groupId = "com.boogiwoogi.di"
+            artifactId = "woogiDi"
+            version = "1.0.0"
+            from(components["kotlin"])
+        }
     }
 }
